+++ src/main/java/com/example/demo/domain/study/StudyOnceComment.java
@@ -53,10 +53,10 @@ public class StudyOnceComment extends BaseEntity {
 
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "study_once_id", foreignKey = @ForeignKey(ConstraintMode.NO_CONSTRAINT))
-	private StudyOnceImpl studyOnce;
+	private StudyOnce studyOnce;
 
 	@Builder
-	private StudyOnceComment(Long id, String content, MemberImpl member, StudyOnceImpl studyOnce) {
+	private StudyOnceComment(Long id, String content, MemberImpl member, StudyOnce studyOnce) {
 		//todo content에 대한 검증 추가
 		this.id = id;
 		this.content = content;
