+++ src/main/java/com/example/demo/repository/study/StudyOnceRepository.java
@@ -4,11 +4,11 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.example.demo.domain.study.StudyOnceImpl;
+import com.example.demo.domain.study.StudyOnce;
 
-public interface StudyOnceRepository extends JpaRepository<StudyOnceImpl, Long>, StudyOnceRepositoryCustom {
+public interface StudyOnceRepository extends JpaRepository<StudyOnce, Long>, StudyOnceRepositoryCustom {
 
 	boolean existsByLeaderId(Long leaderId);
 
-	List<StudyOnceImpl> findByLeaderId(Long leaderId);
+	List<StudyOnce> findByLeaderId(Long leaderId);
 }
