+++ src/main/java/com/example/demo/repository/study/StudyOnceRepositoryCustom.java
@@ -2,11 +2,11 @@ package com.example.demo.repository.study;
 
 import java.util.List;
 
-import com.example.demo.domain.study.StudyOnceImpl;
+import com.example.demo.domain.study.StudyOnce;
 import com.example.demo.dto.study.StudyOnceSearchRequest;
 
 public interface StudyOnceRepositoryCustom {
-	List<StudyOnceImpl> findAllByStudyOnceSearchRequest(StudyOnceSearchRequest studyOnceSearchRequest);
+	List<StudyOnce> findAllByStudyOnceSearchRequest(StudyOnceSearchRequest studyOnceSearchRequest);
 
 	Long count(StudyOnceSearchRequest studyOnceSearchRequest);
 }
